requirements.txt
branca==0.8.2
certifi==2026.2.25
charset-normalizer==3.4.6
folium==0.20.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
urllib3==2.6.3
xyzservices==2025.11.0
